backend/requirements.txt
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.2.1
email-validator==1.2.1
et-xmlfile==1.1.0
fastapi==0.78.0
firebase-admin==5.2.0
Flask==2.1.2
Flask-Cors==3.0.10
flask-expects-json==1.7.0
Flask-JWT-Extended==4.4.2
gcloud==0.18.3
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.2
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.23.1
oauth2client==4.1.3
openai==0.20.0
openpyxl==3.0.10
orjson==3.7.7
pandas==1.4.3
pandas-stubs==1.4.3.220710
Pillow==9.2.0
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tqdm==4.64.0
typing_extensions==4.3.0
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.10
uvicorn==0.17.6
uvloop==0.16.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.0
